Added ability to close tabs, cleaned up class.
There is a known bug where windows cease to open after the initial window is closed.  This is caused by the Window hash in WhisperClient not removing the window from the hash when its closed, hence it attempts to add tabs to a window that isnt visible/exists.  Will fix this evening soon as I get off work...